move code related to task encodings from datamodule to own python module
